+++ src/hooks/users/constraints.ts
@@ -1,0 +1,13 @@
+const API = {
+  CURRENT_USER: {
+    path: "users/me",
+    method: "GET",
+  },
+
+  UPDATE_USER_PROFILE: {
+    path: "users/update",
+    method: "POST",
+  },
+};
+
+export default API;