+++ src/pages/profile/index.tsx
@@ -1,0 +1,3 @@
+export { ProfilePage } from "./profile";
+export { EditProfilePage } from "./edit";
+export { ViewProfile } from "./view";