+++ History.md
@@ -1,5 +1,29 @@
 
-n.n.n / 2022-12-29
+0.0.2 / 2022-12-30
+==================
+
+  * feat: user profile update
+  * feat: user profile details popover and login button
+  * fix: account pop-over links
+  * feat: user profile with populated data
+  * feat: user store and authentication
+  * feat: user profile dashboard view pages
+  * feat: user dashboard details screens
+  * feat: profile card
+  * fix: sign-in ui fixes
+  * chore: added changelog
+  * index title update
+  * fix: small ui issues
+  * feat: home components
+  * feat: header and more theme updates
+  * feat: added theme
+  * fix: booking flow
+  * feat: booking ui
+  * chore: base app migration
+  * chore: config theme
+  * chore: initial commit
+
+0.0.1 / 2022-12-29
 ==================
 
   * index title update